requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.3
googleapis-common-protos==1.70.0
grpcio==1.74.0
idna==3.10
importlib-metadata==8.7.0
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-exporter-otlp-proto-http==1.36.0
opentelemetry-instrumentation==0.57b0
opentelemetry-instrumentation-logging==0.57b0
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
packaging==25.0
protobuf==6.32.0
requests==2.32.5
structlog==25.4.0
typing-extensions==4.14.1
urllib3==2.5.0
wrapt==1.17.3
zipp==3.23.0
